Split node context menu tests
